WIP: back end refactored to handle all denominations. ui renders all the boxes, colors not set yet

+++ app/compiler/Element.scala
@@ -1,3 +1,3 @@
 package compiler
 
-case class Element(original : Boolean, name : String, image : String, color : ElementKinds.Value, value : Int)
+case class Element(original: Boolean, name: String, image: String, color: ElementKinds.Value, value: Double)
